Refactoring add skills
