--- src/features/profile/use-profile-state.ts
@@ -1,11 +1,0 @@
-import { Profile } from "core/profiles/types";
-import { useState } from "react";
-
-export default function useProfileState() {
-  const [profile, setProfile] = useState<Profile | null>(null);
-
-  return {
-    profile,
-    setProfile,
-  };
-}
